Use TOI query for player movement improving precision
Shape-cast the player hitbox for collision with terrain and move it proportional to time-of-impact. Previously I was not moving the player if the new position would result in a terrain intersection. This improved precision and is a form of continuous collision detection. A fast moving player can no longer pass through a thin hitbox. Extracted the player state machine transition function to a method.