Add attack animation, state and keyboard input
Added an attacking state that lasts for the attack animation duration. When the animation is complete the player leaves the attacking state. Added pressed_this_frame to KeyState to check if the attack button was pressed. This prevents the attack from cycling if the attack key is held down and created a more connected feel between controls and in game actions.

Closes #7